@@ -1,5 +1,5 @@
 {
-  "name": "PriceGenie_React_Native",
+  "name": "PriceGenie",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -7,23 +7,35 @@
     "test": "jest"
   },
   "dependencies": {
-    "react": "15.4.2",
-    "react-native": "0.40.0",
+    "axios": "^0.16.1",
+    "lodash": "^4.17.4",
+    "moment": "^2.18.1",
+    "react": "16.0.0-alpha.6",
+    "react-native": "0.44.2",
     "react-native-button": "^1.8.2",
     "react-native-chart": "^1.0.8-beta",
-    "react-native-device-info": "^0.10.1",
-    "react-native-facebook-login": "^1.4.0",
-    "react-native-fcm": "^6.2.0",
-    "react-native-google-signin": "^0.9.0",
-    "react-native-pathjs-charts": "0.0.26",
+    "react-native-deprecated-custom-components": "^0.1.0",
+    "react-native-device-info": "^0.10.2",
+    "react-native-drawer": "^2.3.0",
+    "react-native-facebook-login": "^1.5.0",
+    "react-native-fcm": "^6.2.3",
+    "react-native-google-signin": "^0.10.0",
+    "react-native-pathjs-charts": "0.0.28",
     "react-native-radio-buttons": "^0.14.0",
-    "react-native-vector-icons": "^4.0.0"
+    "react-native-splash-screen": "^2.1.0",
+    "react-native-svg": "^5.1.8",
+    "react-native-vector-icons": "^4.1.1"
   },
   "devDependencies": {
-    "babel-jest": "18.0.0",
-    "babel-preset-react-native": "1.9.1",
-    "jest": "18.1.0",
-    "react-test-renderer": "15.4.2"
+    "babel-jest": "20.0.3",
+    "babel-preset-react-native": "1.9.2",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.0.1",
+    "eslint-plugin-import": "^2.3.0",
+    "eslint-plugin-jsx-a11y": "^5.0.3",
+    "eslint-plugin-react": "^7.0.1",
+    "jest": "20.0.4",
+    "react-test-renderer": "16.0.0-alpha.6"
   },
   "jest": {
     "preset": "react-native"
