@@ -1,15 +1,15 @@
 CONFIG = {
-    baseURL: 'http://pricegenie.co/mobile_api/api.php?action='
-}
+  baseURL: 'http://pricegenie.co/mobile_api/api.php?action=',
+};
 CONFI = {
-    baseURL: 'http://pricegenie.co/price_history_graph.php?work=productgraph&action='
-}
+  baseURL: 'http://pricegenie.co/price_history_graph.php?work=productgraph&action=',
+};
 CONFIGURE = {
-    baseURL: 'http://144.76.34.244:5005/v2/genie_alerts/set_genie_alert/'
-}
+  baseURL: 'http://144.76.34.244:5005/v2/genie_alerts/set_genie_alert/',
+};
 CONF = {
-    baseURL: 'http://pricegenie.co/facebook.php?type='
-}
+  baseURL: 'http://pricegenie.co/facebook.php?type=',
+};
 ADDUSER = {
-    baseURL: 'http://pricegenie.co/mobile_api/api.php?action=add_mobile&device_id='
-}
+  baseURL: 'http://pricegenie.co/mobile_api/api.php?action=add_mobile&device_id=',
+};
