@@ -1,31 +1,33 @@
-export default Buttons_data = [
-    {
-        name: "Price",
-        case: "pricelth",
-        icon: {
-            name: "ios-arrow-round-down-outline",
-            size: 20
-        }
-    }, {
-        name: "Price",
-        case: "pricehtl",
-        icon: {
-            name: "ios-arrow-round-up",
-            size: 20
-        }
-    }, {
-        name: "Popularity",
-        case: "popularity",
-        icon: {
-            name: "ios-heart-outline",
-            size: 0
-        }
-    }, {
-        name: "Filter",
-        case: "filter",
-        icon: {
-            name: "ios-color-filter-outline",
-            size: 20
-        }
-    }
-]
+const Buttons_data = [
+  {
+    name: 'Price',
+    case: 'pricelth',
+    icon: {
+      name: 'ios-arrow-round-down-outline',
+      size: 20,
+    },
+  }, {
+    name: 'Price',
+    case: 'pricehtl',
+    icon: {
+      name: 'ios-arrow-round-up',
+      size: 20,
+    },
+  }, {
+    name: 'Popularity',
+    case: 'popularity',
+    icon: {
+      name: 'ios-heart-outline',
+      size: 0,
+    },
+  },
+  // {
+  //   name: 'Filter',
+  //   case: 'filter',
+  //   icon: {
+  //     name: 'ios-color-filter-outline',
+  //     size: 20,
+  //   },
+  // },
+];
+export default Buttons_data;
