@@ -1,43 +1,43 @@
 export default json_data = [
-    {
-        "case": "MobilesTablets",
-        "image": require('../img/cat_icon/Mobiles.png'),
-        "name": "Mobiles & Tablets ",
-        "pic": "ios-arrow-forward-outline"
-    }, {
-        "case": "CameraAccessories",
-        "image": require('../img/cat_icon/CameraAccessories.png'),
-        "name": "Camera & Accessories ",
-        "pic": "ios-arrow-forward-outline"
-    }, {
-        "case": "LaptopDesktop",
-        "image": require('../img/cat_icon/Laptops.png'),
-        "name": "Laptop & Desktop",
-        "pic": "ios-arrow-forward-outline"
-    }, {
-        "case": "KitchenAppliances",
-        "image": require('../img/cat_icon/KitchenAppliances.png'),
-        "name": "Kitchen Appliances ",
-        "pic": "ios-arrow-forward-outline"
-    }, {
-        "case": "HomeAppliances",
-        "image": require('../img/cat_icon/HomeAppliances.png'),
-        "name": "Home Appliances",
-        "pic": "ios-arrow-forward-outline"
-    }, {
-        "case": "TVHomeAudio",
-        "image": require('../img/cat_icon/Television.png'),
-        "name": "TV & Home Audio",
-        "pic": "ios-arrow-forward-outline"
-    }, {
-        "case": "ComputerAccessories",
-        "image": require('../img/cat_icon/DesktopLaptopAccessories.png'),
-        "name": "Computer Accessories",
-        "pic": "ios-arrow-forward-outline"
-    }, {
-        "case": "ComputerComponents",
-        "image": require('../img/cat_icon/ComputerComponents.png'),
-        "name": "Computer Components",
-        "pic": "ios-arrow-forward-outline"
-    }
+  {
+    case: 'MobilesTablets',
+    image: require('../img/cat_icon/Mobiles.png'),
+    name: 'Mobiles & Tablets ',
+    pic: 'ios-arrow-forward-outline',
+  }, {
+    case: 'CameraAccessories',
+    image: require('../img/cat_icon/CameraAccessories.png'),
+    name: 'Camera & Accessories ',
+    pic: 'ios-arrow-forward-outline',
+  }, {
+    case: 'LaptopDesktop',
+    image: require('../img/cat_icon/Laptops.png'),
+    name: 'Laptop & Desktop',
+    pic: 'ios-arrow-forward-outline',
+  }, {
+    case: 'KitchenAppliances',
+    image: require('../img/cat_icon/KitchenAppliances.png'),
+    name: 'Kitchen Appliances ',
+    pic: 'ios-arrow-forward-outline',
+  }, {
+    case: 'HomeAppliances',
+    image: require('../img/cat_icon/HomeAppliances.png'),
+    name: 'Home Appliances',
+    pic: 'ios-arrow-forward-outline',
+  }, {
+    case: 'TVHomeAudio',
+    image: require('../img/cat_icon/Television.png'),
+    name: 'TV & Home Audio',
+    pic: 'ios-arrow-forward-outline',
+  }, {
+    case: 'ComputerAccessories',
+    image: require('../img/cat_icon/DesktopLaptopAccessories.png'),
+    name: 'Computer Accessories',
+    pic: 'ios-arrow-forward-outline',
+  }, {
+    case: 'ComputerComponents',
+    image: require('../img/cat_icon/ComputerComponents.png'),
+    name: 'Computer Components',
+    pic: 'ios-arrow-forward-outline',
+  },
 ];
